Add some fixes to the scenario
