+++ tests/UI/campaigns/functional/BO/02_orders/01_orders/06_bulkOpenInNewTab.js
