+++ tests/UI/campaigns/functional/BO/02_orders/01_orders/08_viewCustomer.js
